Update dependent code from renames in #86.
Only pkg/amp_validator remains, and that will be moved to
ampproject/amphtml.
